Change jitter implementation (#10)
Decorrelated jitter has a major flaw: clamping. Retry intervals can get
repeatedly clamped to the maximum allowed duration. This effectively
removes any jitter.

The implementation in this library is a variation which exacerbates this
flaw.

This change replaces this jitter implementation with three options:

1. None
2. Full
3. Bounded

It also changes how 'maxiumum retry duration' is implemented. Before,
there was no information about how long the task had been retrying, so
there was no way to implement a correct 'maximum retry duration'.

This has been reimplemented to require a start time for a task when
setting a maxiumum retry duration.

* Fix doc reference
* Improve changelog